refactor: move properties into profile package
