+++ integration/osrmfiles/osrmtype/profile/testdata/README.md
@@ -1,0 +1,4 @@
+# testdata
+
+## properties
+It's `/common/properties` which comes from compiled OSRM file `.osrm.properties`, can be extract by `tar -xvf .osrm.properties`.
